[TUBEMQ-567] broker version control

+++ tubemq-manager/src/main/java/org/apache/tubemq/manager/controller/node/request/CloneBrokersReq.java
@@ -30,5 +30,5 @@ public class CloneBrokersReq {
     private List<AddTopicReq> addTopicReqs;
     private String confModAuthToken;
     private String createUser;
-    private int clusterId;
+    private Long clusterId;
 }
